refactor: split omnitable settings
