+++ lib/settings/context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'haunted';
+
+export const SettingsContext = createContext();
+
+customElements.define('settings-provider', SettingsContext.Provider);
